requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.15
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
mongoengine==0.29.1
ordered-set==4.1.0
packaging==24.2
Pygments==2.18.0
PyJWT==2.10.0
pymongo==4.10.1
python-dotenv==1.0.1
redis==5.2.0
rich==13.9.4
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.17.0
